+++ client/landing/stepper/declarative-flow/internals/steps-repository/assign-trial-plan/constants.ts
@@ -1,0 +1,4 @@
+export enum AssignTrialResult {
+	SUCCESS = 'success',
+	FAILURE = 'failure',
+}
