+++ client/landing/stepper/declarative-flow/internals/steps-repository/processing-step/constants.ts
@@ -1,0 +1,5 @@
+export enum ProcessingResult {
+	NO_ACTION = 'no-action',
+	SUCCESS = 'success',
+	FAILURE = 'failure',
+}
