+++ client/landing/stepper/declarative-flow/site-setup-flow.ts
@@ -42,7 +42,8 @@ import ImporterWix from './internals/steps-repository/importer-wix';
 import ImporterWordpress from './internals/steps-repository/importer-wordpress';
 import IntentStep from './internals/steps-repository/intent-step';
 import PatternAssembler from './internals/steps-repository/pattern-assembler/lazy';
-import ProcessingStep, { ProcessingResult } from './internals/steps-repository/processing-step';
+import ProcessingStep from './internals/steps-repository/processing-step';
+import { ProcessingResult } from './internals/steps-repository/processing-step/constants';
 import SiteOptions from './internals/steps-repository/site-options';
 import SiteVertical from './internals/steps-repository/site-vertical';
 import StoreAddress from './internals/steps-repository/store-address';
